Move IncidentFilter to separate file
